# specs/001-research-data-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Research Data Collection Platform API
  description: |
    API for the ICU Delirium Psychoeducational Study platform.
    Supports pre/post-test questionnaires, educational materials, evaluator scoring, and data export.
  version: 1.0.0
  contact:
    name: Research Platform Team
servers:
  - url: /api/v1
    description: API base path
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      description: Uses HttpOnly refresh token cookie
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out successfully
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change current user password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          description: Validation error

  # ============ STUDENT ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/UserInfo'
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, evaluator, student]

    # Student
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string